server/api/requirements.txt
flask
pillow
numpy
tensorflow
matplotlib
